Move DirectJoin to separate file
